src/pyproject.toml
[project]
name = "local-auctions-python-scraper"
version = "1.0.0"
description = "A tool for scraping localauctions.com since their website sucks ass"
authors = [
    {name = "Nick Melillo",email = "vyse323@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "pycurl (>=7.45.6,<8.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
# Ensures we can run `poetry sync` without obliterating poetry in the process
poetry = "^2.1.4"
poetry-plugin-sort = "^0.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
